Improve websocket handling logic

+++ server/controllers/ws_controller.py
@@ -13,6 +13,19 @@ class WebSocketController(SessionMixin, WebSocketHandler):
 
     clients = []
 
+    def update_session(self):
+        with self.make_session() as session:
+            entry = session.get(Session, self.request.remote_ip)
+            if entry:
+                entry.last_ping = self.ws_connection.last_ping
+                entry.last_pong = self.ws_connection.last_pong
+                session.commit()
+            else:
+                session.add(Session(remote_ip=self.request.remote_ip,
+                                    last_ping=self.ws_connection.last_ping,
+                                    last_pong=self.ws_connection.last_pong))
+                session.commit()
+
     def data_received(self, chunk: bytes) -> Optional[Awaitable[None]]:
         raise RuntimeError(
             f'Data streaming not supported for {self.__class__}')
@@ -27,12 +40,7 @@ class WebSocketController(SessionMixin, WebSocketHandler):
 
         # TODO: This assumes only one session from a single client IP, which
         # might not be true
-        with self.make_session() as session:
-            session.add(Session(remote_ip=self.request.remote_ip,
-                                last_ping=self.ws_connection.last_ping,
-                                last_pong=self.ws_connection.last_pong))
-            session.commit()
-
+        self.update_session()
         get_logger().info(f'WebSocket opened from: {self.request.remote_ip}')
 
     def on_close(self) -> None:
@@ -43,21 +51,15 @@ class WebSocketController(SessionMixin, WebSocketHandler):
         # might not be true
         with self.make_session() as session:
             entry = session.get(Session, self.request.remote_ip)
-            session.delete(entry)
-            session.commit()
+            if entry:
+                session.delete(entry)
+                session.commit()
 
         get_logger().info(f'WebSocket closed from: {self.request.remote_ip}')
 
     def on_message(self, message: Union[str, bytes]
                    ) -> Optional[Awaitable[None]]:
         pass
-
-    def update_session(self):
-        with self.make_session() as session:
-            entry = session.get(Session, self.request.remote_ip)
-            entry.last_ping = self.ws_connection.last_ping
-            entry.last_pong = self.ws_connection.last_pong
-            session.commit()
 
     def on_pong(self, data: bytes) -> None:
         self.update_session()
